algorand-protocol/py-sdk/requirements.txt
cffi==1.15.1
msgpack==1.0.5
py-algorand-sdk==2.1.2
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.5.0
python-dotenv==1.0.0
